+++ cmd/goluca/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"context"
 	"log"
+	"net/http"
 
 	"github.com/hampgoodwin/GoLuca/internal/environment"
 	"go.uber.org/zap"
@@ -15,7 +16,12 @@ func main() {
 		log.Panic("failed to create new environment")
 	}
 
-	s := environment.NewHTTPServer(env)
+	// s := environment.NewHTTPServer(env)
+	s := &http.Server{
+		Addr:     env.Config.HTTPAPI.AddressString(),
+		ErrorLog: zap.NewStdLog(env.Log),
+		Handler:  env.HTTPMux,
+	}
 
 	if err := s.ListenAndServe(); err != nil {
 		env.Log.Panic("http server failed", zap.Error(err))
